backend/requirements.txt
mysqlclient
mysql-connector-python
djangorestframework
django-livereload-server